* Reordering files location to remove any trace of progressbar   from the transpiler component.
